change `spos_t`

+++ libufold/include/ufold/Separators.hpp
@@ -26,13 +26,14 @@
 #ifndef UFOLD_SEPARATORS_HPP
 #define UFOLD_SEPARATORS_HPP
 
+#include <cstdint>
 #include <map>
 
 /// Namespace of libufold
 namespace ufold
 {
 
-    using spos_t = uint8_t; ///< The type representing line column
+    using spos_t = uint16_t; ///< The type representing line column
     using separator_t = uint8_t; ///< The underlying type of #SeparatorType
 
     /// `enum` of different type of separators
